# mental-health/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI 心理健康助手 API
  description: |
    鸿蒙端 + Web 端共用接口。
    - 鉴权方式：JWT，请求头 `Authorization: Bearer <token>`
    - 文件上传：multipart/form-data
    - AI 流式：SSE（text/event-stream）
    - 实时消息：WebSocket `wss://39.96.36.94/ws/im?token=xxx`
  version: 1.0.0
  contact:
    name: hwx-1
servers:
  - url: http://39.96.36.94:8080/api
    description: 生产环境
  - url: http://localhost:8080/api
    description: 本地开发
...
paths:
...
  /ai/sessions:
    get:
      tags: [AI]
      summary: 会话列表
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: business, in: query, schema: { type: string, enum: [mental, study, general] } }
      responses:
        '200': { description: OK }
    post:
      tags: [AI]
      summary: 创建会话
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [business]
              properties:
                business: { type: string, enum: [mental, study, general] }
                title: { type: string }
      responses:
        '200': { description: OK }
...
  /ai/sessions/{id}/messages:
    get:
      tags: [AI]
      summary: 会话历史消息
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...